requirements.txt
streamlit
pandas
matplotlib
reportlab
Pillow
python-docx
